separate reusable two-way-auth functions out of route
