+++ apps/next/src/app/two-way-auth/api/types.ts
@@ -1,0 +1,5 @@
+export interface Factor {
+	value: string
+	expires: number
+	lifespan: number
+}
